services/subtitle-generator/requirements.txt
gunicorn==21.2.0
opencv-python-headless==4.8.0.76
flask-cors==4.0.0
pika==1.3.2
Flask==3.0.0
flask-debug==0.4.3
watchdog==3.0.0
celery==5.3.6
beautifulsoup4==4.12.3
protobuf==5.29.3
kombu==5.4.2
boto3==1.36.14
python-dotenv==1.0.1
ffmpeg-python==0.2.0
openai-whisper==20240930
